+++ Makefile.nmake
@@ -6,12 +6,13 @@ TINYCBOR_SOURCES = \
 	src\cborencoder.c \
 	src\cborencoder_close_container_checked.c \
 	src\cborparser.c \
+	src\cborparser_dup_string.c \
 	src\cborpretty.c
 TINYCBOR_OBJS = \
 	src\cborerrorstrings.obj \
 	src\cborencoder.obj \
 	src\cborencoder_close_container_checked.obj \
-	src\cborparser.obj \
+	src\cborparser_dup_string.obj \
 	src\cborpretty.obj
 
 all: lib\tinycbor.lib
